small fixes

+++ Template/NLO/SubProcesses/handling_lhe_events.f
@@ -671,7 +671,7 @@ c
          elseif(jwgtinfo.eq.-5) then
            write(ifile,'(a)')'  <rwgt>'
            write (ifile,'(1x,d16.10,3(1x,i4))') wgtref,n_ctr_found
-     &          ,n_mom_conf,wgtcpower
+     &          ,n_mom_conf,nint(wgtcpower)
            do i=1,n_mom_conf
               do j=1,mexternal
                  write (ifile,'(4(1x,d16.10))')
